fix: devided recordings and events to live components
